Data Loading refactored
