+++ trajnetbaselines/lstm/data_load_utils.py
@@ -1,0 +1,58 @@
+import trajnetplusplustools
+import os
+import pickle
+
+def prepare_data(path, subset='/train/', sample=1.0, goals=True):
+    """ Prepares the train/val scenes and corresponding goals 
+    
+    Parameters
+    ----------
+    subset: String ['/train/', '/val/']
+        Determines the subset of data to be processed
+    sample: Float (0.0, 1.0]
+        Determines the ratio of data to be sampled
+    goals: Bool
+        If true, the goals of each track are extracted
+        The corresponding goal file must be present in the 'goal_files' folder
+        The name of the goal file must be the same as the name of the training file
+
+    Returns
+    -------
+    all_scenes: List
+        List of all processed scenes
+    all_goals: Dictionary
+        Dictionary of goals corresponding to each dataset file.
+        None if 'goals' argument is False.
+    Flag: Bool
+        True if the corresponding folder exists else False.
+    """
+
+    ## Check if folder exists
+    if not os.path.isdir(path + subset):
+        if 'train' in subset:
+            print("Train folder does NOT exist")
+            exit()
+        if 'val' in subset:
+            print("Validation folder does NOT exist")
+            return None, None, False
+
+    ## read goal files
+    all_goals = {}
+    all_scenes = []
+
+    ## List file names
+    files = [f.split('.')[-2] for f in os.listdir(path + subset) if f.endswith('.ndjson')]
+    ## Iterate over file names
+    for file in files:
+        reader = trajnetplusplustools.Reader(path + subset + file + '.ndjson', scene_type='paths')
+        ## Necessary modification of train scene to add filename
+        scene = [(file, s_id, s) for s_id, s in reader.scenes(sample=sample)]
+        if goals:
+            goal_dict = pickle.load(open('goal_files/' + subset + file +'.pkl', "rb"))
+            ## Get goals corresponding to train scene
+            all_goals[file] = {s_id: [goal_dict[path[0].pedestrian] for path in s] for _, s_id, s in scene}
+        all_scenes += scene
+
+    if goals:
+        return all_scenes, all_goals, True
+    return all_scenes, None, True
